moving last unsorted component
