+++ frontend/app/[locale]/methodology/_components/animated-afflictions.tsx
